Refactored Player/ Enemy modules to separate files
