+++ src/utils/models/enemy/mod.rs
@@ -1,0 +1,4 @@
+pub struct Enemy {
+  pub max_health: i32,
+  pub current_health: i32,
+}
